Upload Image + tag

+++ ElevatorSystem.Admin/ElevatorSystem.Admin/Controllers/AdminController/CategoriesController.cs
@@ -102,6 +102,7 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Category category = db.Categories.Find(id);
+           
             if (category == null)
             {
                 return HttpNotFound();
@@ -117,10 +118,31 @@ namespace ElevatorSystem.Admin.Controllers.AdminController
            
             Category category = db.Categories.Find(id);
             category.DeletedAt = DateTime.Today;
-            db.Categories.Remove(category);
-            db.SaveChanges();
-            TempData["DeleteMessage"] = "Category { #" + category.ID + "." + category.Name + " } has been removed from the list !";
-            return RedirectToAction("Index");
+
+            //Check Product List
+            IEnumerable<Elevator> elevatorList = db.Elevators.ToList();
+            var result = elevatorList.GroupBy(x => x.Category)
+                                     .Where(x => x.Count() >= 1)
+                                     .Select(y => new { category = y.Key, Count = y.Count() });
+
+            var result1 = from product1 in db.Elevators
+                         where (product1.CategoryID == id)
+                          select new { product1};
+
+            if(result1.Count() < 1)
+            {
+                db.Categories.Remove(category);
+                db.SaveChanges();
+                TempData["DeleteMessage"] = "Category { #" + category.ID + "." + category.Name + " } has been removed from the list !";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                ViewData["InvalidRemove"] = "Cannot remove category { #" + category.ID + "." + category.Name + " } because it has ";
+                return View(category);
+            }
+            
+         
         }
 
         protected override void Dispose(bool disposing)
